fix(core): converge Work History lifecycle ownership
