--- framework/core/src/python/kungfu/workspace_history/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-"""Work History lifecycle ownership namespace."""
